Building Multimodal Search Engine/requirements.txt
torch
clip-by-openai
pillow
faiss_gpu
numpy
streamlit
tqdm